Moved Message board in beamcut_CASA_calibration.py to utils/user_interaction.py
